+++ dffml/model/model.py
@@ -79,6 +79,14 @@ class Model(BaseDataFlowFacilitatorObject):
 
     CONFIG = ModelConfig
 
+    def __init__(self, config):
+        super().__init__(config)
+        # TODO Just in case its a string. We should make it so that on
+        # instantiation of an @config we convert properties to their correct
+        # types.
+        if isinstance(getattr(self.config, "directory", None), str):
+            self.config.directory = pathlib.Path(self.config.directory)
+
     def __call__(self) -> ModelContext:
         self._make_config_directory()
         return self.CONTEXT(self)
@@ -197,16 +205,20 @@ class SimpleModel(Model):
 
     def check_applicable_feature(self, feature):
         # Check the data datatype is in the list of supported data types
-        if feature.dtype() not in self.DTYPES:
+        self.check_feature_dtype(feature.dtype())
+        # Check that length (dimensions) of feature is supported
+        self.check_feature_length(feature.length())
+        return True
+
+    def check_feature_dtype(self, dtype):
+        if dtype not in self.DTYPES:
             msg = f"{self.__class__.__qualname__} only supports features "
             msg += f"with these data types: {self.DTYPES}"
             raise ValueError(msg)
+
+    def check_feature_length(self, length):
         # If SUPPORTED_LENGTHS is None then all lengths are supported
-        if self.SUPPORTED_LENGTHS is None:
-            return True
-        # Check that length (dimensions) of feature is supported
-        if feature.length() not in self.SUPPORTED_LENGTHS:
+        if self.SUPPORTED_LENGTHS and length not in self.SUPPORTED_LENGTHS:
             msg = f"{self.__class__.__qualname__} only supports "
             msg += f"{self.SUPPORTED_LENGTHS} dimensional values"
             raise ValueError(msg)
-        return True