+++ model/tensorflow/dffml_model_tensorflow/dnnr.py
@@ -3,6 +3,7 @@ Uses Tensorflow to create a generic DNN which learns on all of the features in a
 record.
 """
 import os
+import pathlib
 from typing import List, Dict, Any, AsyncIterator
 
 import numpy as np
@@ -29,11 +30,9 @@ class DNNRegressionModelConfig:
     epochs: int = field(
         "Number of iterations to pass over all records in a source", default=30
     )
-    directory: str = field(
+    directory: pathlib.Path = field(
         "Directory where state should be saved",
-        default=os.path.join(
-            os.path.expanduser("~"), ".cache", "dffml", "tensorflow"
-        ),
+        default=pathlib.Path("~", ".cache", "dffml", "tensorflow"),
     )
     hidden: List[int] = field(
         "List length is the number of hidden layers in the network. Each entry in the list is the number of nodes in that hidden layer",
@@ -196,58 +195,32 @@ class DNNRegressionModel(Model):
       make sure to take a BACKUP of files with same name in the directory
       from where this command is run as it overwrites any existing files.
 
-    .. code-block:: console
+    .. literalinclude:: /../model/tensorflow/examples/tfdnnr/train_data.sh
 
-        $ cat > train.csv << EOF
-        Feature1,Feature2,TARGET
-        0.93,0.68,3.89
-        0.24,0.42,1.75
-        0.36,0.68,2.75
-        0.53,0.31,2.00
-        0.29,0.25,1.32
-        0.29,0.52,2.14
-        EOF
-        $ cat > test.csv << EOF
-        Feature1,Feature2,TARGET
-        0.57,0.84,3.65
-        0.95,0.19,2.46
-        0.23,0.15,0.93
-        EOF
-        $ dffml train \\
-            -model tfdnnr \\
-            -model-epochs 300 \\
-            -model-steps 2000 \\
-            -model-predict TARGET:float:1 \\
-            -model-hidden 8 16 8 \\
-            -sources s=csv \\
-            -source-filename train.csv \\
-            -model-features \\
-              Feature1:float:1 \\
-              Feature2:float:1 \\
-            -log debug
-        Enabling debug log shows tensorflow losses...
-        $ dffml accuracy \\
-            -model tfdnnr \\
-            -model-predict TARGET:float:1 \\
-            -model-hidden 8 16 8 \\
-            -sources s=csv \\
-            -source-filename test.csv \\
-            -model-features \\
-              Feature1:float:1 \\
-              Feature2:float:1 \\
-            -log critical
+    .. literalinclude:: /../model/tensorflow/examples/tfdnnr/test_data.sh
+
+    Train the model
+
+    .. literalinclude:: /../model/tensorflow/examples/tfdnnr/train.sh
+
+    Assess the accuracy
+
+    .. literalinclude:: /../model/tensorflow/examples/tfdnnr/accuracy.sh
+
+    Output
+
+    .. code-block::
+
         0.9468210011
-        $ echo -e 'Feature1,Feature2,TARGET\\n0.21,0.18,0.84\\n' | \\
-          dffml predict all \\
-            -model tfdnnr \\
-            -model-predict TARGET:float:1 \\
-            -model-hidden 8 16 8 \\
-            -sources s=csv \\
-            -source-filename /dev/stdin \\
-            -model-features \\
-              Feature1:float:1 \\
-              Feature2:float:1 \\
-            -log critical
+
+    Make a prediction
+
+    .. literalinclude:: /../model/tensorflow/examples/tfdnnr/predict.sh
+
+    Output
+
+    .. code-block:: json
+
         [
             {
                 "extra": {},
@@ -266,6 +239,10 @@ class DNNRegressionModel(Model):
                 "key": 0
             }
         ]
+
+    Example usage of Tensorflow DNNEstimator model using python API
+
+    .. literalinclude:: /../model/tensorflow/examples/tfdnnr/tfdnnr.py
 
     The ``NaN`` in ``confidence`` is the expected behaviour. (See TODO in
     predict).