+++ tests/operation/test_dataflow.py
@@ -1,3 +1,5 @@
+import copy
+
 from dffml.df.types import DataFlow, Input
 from dffml.df.memory import MemoryOrchestrator
 from dffml.operation.dataflow import run_dataflow, RunDataFlowConfig
@@ -75,5 +77,48 @@ class TestRunDataFlowOnRecord(AsyncTestCase):
                     self.assertIn("result", results)
 
                     results = results["result"]
+                    expected_results = test_outputs[ctx_str]
+                    self.assertEqual(expected_results, results)
+
+    async def test_run_custom(self):
+        output_definition = add.op.outputs["sum"]
+
+        get_single_spec_input = Input(
+            value=[output_definition.name],
+            definition=GetSingle.op.inputs["spec"],
+        )
+
+        subflow = copy.deepcopy(DATAFLOW)
+        subflow.seed.append(get_single_spec_input)
+
+        test_dataflow = DataFlow(
+            operations={
+                "run_dataflow": run_dataflow.op._replace(
+                    inputs=parse_line.op.inputs,
+                    outputs={output_definition.name: output_definition},
+                ),
+                "get_single": GetSingle.imp.op,
+            },
+            configs={"run_dataflow": RunDataFlowConfig(dataflow=subflow)},
+            seed=[get_single_spec_input],
+        )
+
+        test_outputs = {"add 40 and 2": 42, "multiply 42 and 10": 420}
+
+        async with MemoryOrchestrator.withconfig({}) as orchestrator:
+            async with orchestrator(test_dataflow) as octx:
+                async for _ctx, results in octx.run(
+                    {
+                        input_line: [
+                            Input(
+                                value=input_line,
+                                definition=parse_line.op.inputs["line"],
+                            )
+                        ]
+                        for input_line in test_outputs
+                    }
+                ):
+                    ctx_str = (await _ctx.handle()).as_string()
+                    results = results[output_definition.name]
                     expected_results = test_outputs[ctx_str]
                     self.assertEqual(expected_results, results)